pyproject.toml
[tool.poetry]
name = "raspicam"
version = "0.1.0"
description = "Code for the MHP camera system"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
opencv-python = "^4"
paho-mqtt = "^1.5.0"
python-dotenv = "^0.13.0"
# Raspberry Pi uses the ARM platform, so these will only install on a Pi
picamera = { version = "^1.13", markers = "platform_machine == 'armv6l' or platform_machine == 'armv7l'" }
"RPi.GPIO" = { version = "^0.7.0", markers = "platform_machine == 'armv6l' or platform_machine == 'armv7l'" }
adafruit-circuitpython-mcp3xxx = { version = "^1.4.5", markers = "platform_machine == 'armv6l' or platform_machine == 'armv7l'" }
mhp = { git = "git@github.com:monash-human-power/common.git", rev = "202212.21" }
python-can = "^4.2.2"

[tool.poetry.dev-dependencies]
pylint = "^2.5.2"
pytest = "^5.4.1"
black = "^19.10b0"
flake8 = "^3.8.3"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[pycalver]
current_version = "v202009.0001-alpha"
version_pattern = "{pycalver}"
commit = true
tag = true
push = true

[pycalver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"']
"README.md" = ["{version}", "{pep440_version}"]
